refactor

+++ src/NeuralNetworkRandomFactory.go
@@ -30,7 +30,7 @@ func (factory *NeuralNetworkRandomFactory) GenerateFloat(min float64, max float6
 func (factory *NeuralNetworkRandomFactory) Make() NeuralNetwork {
 	var size = ComputeNumberOfWeights()
 	var weights = make([]float64, size)
-	for i := uint(0); i < size; i++ {
+	for i := 0; i < size; i++ {
 		weights[i] = factory.GenerateFloat(MIN_RAND_WEIGHT, MAX_RAND_WEIGHT)
 	}
 	return NeuralNetwork{weights}
